requirements-test.txt
pytest>=7.4
pytest-django>=4.7
responses>=0.25
